+++ macos/Classes/FlutterRTCPeerConnection.h
@@ -1,1 +1,0 @@
-../../common/darwin/Classes/FlutterRTCPeerConnection.h